requirements.txt
Django==4.2.7
gunicorn>=20.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
dj-database-url>=2.1.0
setuptools>=65.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow>=10.4.0
docker==5.0.3
requests==2.31.0
urllib3==1.26.18
